Fix incorrect usage of `unsigned` and `int` in place of NSUInteger/NSInteger's.
On x64, NSInteger is 8 bytes long while `int` is only 4. This causes problems when ex. NSNotFound
is cast to an int and then compared to itself.

There probably are many more places the integer types are used the wrong way, and this is likely to
cause us more headache later.