Get rid of the window/backing context types discrepancy
From here on, all the contexts are "backing"; that is, they reside somewhere in the memory
(RAM, GPU memory, files, etc), and the windowing backend is responsible for displaying the
contents on the screen somehow. The way they actually do that is by sending the pixel data
over to the GPU, composing the final image using OpenGL rendering and displaying it using
OpenGL.framework.