Implement basic NSSavePanel & NSOpenPanel-s directly in Cocoa
The upstream Cocotron assumes that the implementation of the file panels
is provided by the display, i.e. by the native toolkit. That is what we
ultimately want as well, but we're far from being there yet, and we also
don't want to make any toolkit a hard requirement.

So in order to have our cake and eat it too, these panel implementations
are designed work entirely inside Cocoa but also let the display provide
a native implementation instead if one is available.